Improved bloom and depth of field

+++ code/shaders/bloom/brightFilter.frag
@@ -10,7 +10,7 @@ uniform sampler2D tex;
 void main()
 {
 	vec2 NDC = ((FragPos.xy + 1.0) * .5);
-	int degree = 3;
+	float degree = 3.0;
 	vec4 color = texture(tex, NDC);
 	float brightness = (color.r * 0.2126) + (color.g * 0.7152) + (color.b * 0.0722);
 	outColor = color * pow(brightness, degree);
